requirements.txt
asgiref==3.9.1
distlib==0.4.0
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
gunicorn